Integrate Bevy game engine into main function
The main function has been updated to use the Bevy game engine. The
previous 'Hello, world!' print statement has been replaced with the
initialization of a new Bevy application. The DefaultPlugins have been
added to the application to provide basic functionality. The application
is then run to start the game engine.